Move TrapFocus to separate Unstable_TrapFocus folder
