+++ packages/material-ui/src/Modal/index.d.ts
@@ -2,5 +2,3 @@ export { default } from './Modal';
 export * from './Modal';
 export { default as ModalManager } from './ModalManager';
 export * from './ModalManager';
-export { default as TrapFocus } from './TrapFocus';
-export * from './TrapFocus';
