+++ packages/material-ui/src/Unstable_TrapFocus/TrapFocus.d.ts
