+++ packages/material-ui/src/Unstable_TrapFocus/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from '../Unstable_TrapFocus/TrapFocus';
+export * from '../Unstable_TrapFocus/TrapFocus';
