+++ packages/material-ui/src/Unstable_TrapFocus/index.js
@@ -1,0 +1,1 @@
+export { default } from "./TrapFocus";
