+++ packages/material-ui/src/index.js
@@ -339,6 +339,10 @@ export * from './Tooltip';
 export { default as Typography } from './Typography';
 export * from './Typography';
 
+// eslint-disable-next-line camelcase
+export { default as Unstable_TrapFocus } from './Unstable_TrapFocus';
+export * from './Unstable_TrapFocus';
+
 export { default as useMediaQuery } from './useMediaQuery';
 export * from './useMediaQuery';
